requirements.txt
azure-storage-blob>=12.16.0
yfinance>=0.2.0
pandas>=1.4.0
matplotlib>=3.5.0
